update: react18, typescript applied
