+++ .env.sample
@@ -1,3 +1,3 @@
-FAUCET_MNEMONIC=
-RECAPTCHA_SITEKEY=
-EXPLORER_URL=
+REACT_APP_FAUCET_MNEMONIC=
+REACT_APP_RECAPTCHA_SITEKEY=
+REACT_APP_EXPLORER_URL=
