+++ .gitignore
@@ -28,4 +28,4 @@ yarn-error.log*
 yarn.lock
 package-lock.json
 
-src/config.js
+src/config.*
