+++ .prettierrc
@@ -1,0 +1,7 @@
+{
+  "semi": true,
+  "singleQuote": true,
+  "printWidth": 150,
+  "useTabs": true,
+  "tabWidth": 4
+}
