+++ package.json
@@ -1,56 +1,76 @@
 {
-  "name": "firma-faucet",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@cosmjs/crypto": "^0.26.0",
-    "@cosmjs/launchpad": "^0.26.0",
-    "@cosmjs/proto-signing": "^0.26.0",
-    "@cosmjs/stargate": "^0.26.0",
-    "@firmachain/firma-js": "^0.2.22",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "copy-to-clipboard": "^3.3.1",
-    "ipfs-http-client": "^52.0.3",
-    "react": "^17.0.2",
-    "react-app-env": "^1.2.3",
-    "react-dom": "^17.0.2",
-    "react-google-recaptcha": "^2.1.0",
-    "react-redux": "^7.2.4",
-    "react-scripts": "4.0.3",
-    "redux-devtools-extension": "^2.13.9",
-    "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.3.0",
-    "styled-components": "^5.3.1",
-    "typescript": "^4.4.2",
-    "web-vitals": "^1.0.1"
-  },
-  "scripts": {
-    "start-testnet": "react-app-env --env-file=.env.testnet start",
-    "build-testnet": "GENERATE_SOURCEMAP=false react-app-env --env-file=.env.testnet build",
-    "deploy-testnet": "aws s3 sync ./build s3://deploy-ecosystem/testnet/faucet --profile=firmachain-deploy --delete",
-    "invalidate-testnet": "aws cloudfront create-invalidation --profile=firmachain-deploy --distribution-id E3KRC3IQXC40MN --paths '/*'"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+	"name": "firma-faucet",
+	"version": "0.1.0",
+	"private": true,
+	"dependencies": {
+		"@cosmjs/crypto": "^0.26.0",
+		"@cosmjs/launchpad": "^0.26.0",
+		"@cosmjs/proto-signing": "^0.26.0",
+		"@cosmjs/stargate": "^0.26.0",
+		"@firmachain/firma-js": "^0.2.62",
+		"@material-ui/core": "^4.12.3",
+		"@material-ui/icons": "^4.11.2",
+		"@material-ui/lab": "^4.0.0-alpha.60",
+		"axios": "^1.7.7",
+		"copy-to-clipboard": "^3.3.1",
+		"ipfs-http-client": "^52.0.3",
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1",
+		"react-google-recaptcha": "^2.1.0",
+		"react-redux": "^7.2.4",
+		"react-scripts": "^5.0.1",
+		"redux-devtools-extension": "^2.13.9",
+		"redux-persist": "^6.0.0",
+		"redux-thunk": "^2.3.0",
+		"styled-components": "^5.3.1",
+		"web-vitals": "^1.0.1"
+	},
+	"scripts": {
+		"start-testnet": "env-cmd -f .env.testnet craco start",
+		"build-testnet": "GENERATE_SOURCEMAP=false env-cmd -f .env.testnet craco build",
+		"deploy-testnet": "aws s3 sync ./build s3://deploy-ecosystem/testnet/faucet --profile=firmachain-deploy --delete",
+		"invalidate-testnet": "aws cloudfront create-invalidation --profile=firmachain-deploy --distribution-id E3KRC3IQXC40MN --paths '/*'",
+		"format": "prettier --write \"src/**/*.{ts,tsx}\""
+	},
+	"eslintConfig": {
+		"extends": [
+			"react-app",
+			"react-app/jest"
+		]
+	},
+	"browserslist": {
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
+	},
+	"devDependencies": {
+		"@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+		"@craco/craco": "^7.1.0",
+		"@testing-library/jest-dom": "^5.11.4",
+		"@testing-library/react": "^11.1.0",
+		"@testing-library/user-event": "^12.1.10",
+		"@types/node": "^22.8.7",
+		"@types/react": "^17.0.83",
+		"@types/react-dom": "^18.3.1",
+		"@types/react-google-recaptcha": "^2.1.9",
+		"@types/styled-components": "^5.1.34",
+		"ajv": "^7.2.4",
+		"browserify-fs": "^1.0.0",
+		"crypto-browserify": "^3.12.1",
+		"env-cmd": "^10.1.0",
+		"path": "^0.12.7",
+		"prettier": "^3.3.3",
+		"process": "^0.11.10",
+		"stream-browserify": "^3.0.0",
+		"typescript": "^5.6.3",
+		"vm-browserify": "^1.1.2",
+		"webpack": "^5.95.0"
+	}
 }
