--- src/App.js
@@ -1,12 +1,0 @@
-import Main from "./screen/main"
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <Main />
-    </div>
-  );
-}
-
-export default App;
